+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 12/12/2023 - v4.0.5
+
+- (Add) Tool - Phased Exposure: Replace "Double Exposure" with "Phased Exposure" tool, same effect but allow to define more steps (#778)
+- (Fix) Infill: Honeycomb Infill does not completely infill model (#789)
+- (Improvement) File -> Open recent: Disable all inexistent files (Can no longer be clicked)
+- (Upgrade) .NET from 6.0.24 to 6.0.25
+- (Upgrade) AvaloniaUI from 11.0.5 to 11.0.6
+- 
 ## 09/11/2023 - v4.0.4
 
 - **FileFormat Anycubic:**